+++ src/app.module.ts
@@ -7,7 +7,7 @@ import { ConfigModule } from '@nestjs/config';
 import { JwtModule } from '@nestjs/jwt';
 
 import { UserModule } from './user/user.module';
-import { AuthModule } from './auth/auth.module';
+// import { AuthModule } from './auth/auth.module';
 import { FeedModule } from './feed/feed.module';
 import { ClothesModule } from './clothes/clothes.module';
 import CatchException from './utils/CatchException';
@@ -25,13 +25,15 @@ import CatchException from './utils/CatchException';
       signOptions: { expiresIn: '12h' },
     }),
     UserModule,
-    AuthModule,
+    // AuthModule,
     FeedModule,
     ClothesModule,
   ],
-  providers: [{
-    provide: APP_FILTER,
-    useClass: CatchException,
-  }]
+  providers: [
+    {
+      provide: APP_FILTER,
+      useClass: CatchException,
+    },
+  ],
 })
 export class AppModule {}
