+++ src/user/user.controller.ts
@@ -14,9 +14,12 @@ import {
   Res,
   Delete,
   Put,
+  Query,
 } from '@nestjs/common';
 import { Request } from 'express';
 import { JwtService } from '@nestjs/jwt';
+import { ConfigService } from '@nestjs/config';
+import { ApiCookieAuth } from '@nestjs/swagger';
 
 import { UserService } from './user.service';
 import {
@@ -29,6 +32,7 @@ import { AuthGuard } from '@nestjs/passport';
 import { TokenService } from 'src/utils/verifyToken';
 import { UserFollowEntity } from 'src/entities/userFollows.entity';
 import { UserEntity } from 'src/entities/users.entity';
+// import { AuthService } from 'src/auth/auth.service';
 
 // 회원가입 : 회원가입 상세, 로그아웃, 회원탈퇴, 회원 정보 수정(O), 닉네임 중복 체크(O)
 // 유저 팔로우 : 목록(O), 생성(O), 삭제(O)
@@ -40,9 +44,50 @@ export class UserController {
     private readonly userService: UserService,
     private readonly JwtService: JwtService,
     private readonly tokenService: TokenService,
+    // private readonly authService: AuthService,
+    private configService: ConfigService,
   ) {}
 
+  // -----소셜로그인 : auth → user_ing
+  @ApiCookieAuth() //  Swagger (OpenAPI) 문서를 자동으로 생성, Swagger 문서에서 특정 API 엔드포인트가 쿠키 기반의 인증을 사용한다는 정보를 표시
+  @Get('/kakao/callback') // 카카오 서버를 거쳐서 도착하게 될 엔드포인트
+  @UseGuards(AuthGuard('kakao')) // kakao.strategy를 실행시켜 준다.
+  @HttpCode(301)
+  async kakaoLogin(
+    @Req() req: Request,
+    // @Res() res: Response,
+    @Query('code') code: string,
+  ): Promise<LoginResponseDto> {
+    console.log(`callback  ///  code------------ ${code}`);
+
+    const { token } = await this.userService.getJWT(
+      req.user.kakaoId,
+      req.user.kakaoEmail,
+      req.user.kakaoNickname,
+      req.user.kakaoProfileImage,
+    );
+
+    // 쿠키 여부 확인 필요
+    // res.cookie('accessToken', accessToken, { httpOnly: true });
+    // res.cookie('refreshToken', refreshToken, { httpOnly: true });
+    // res.cookie('isLoggedIn', true, { httpOnly: false });
+
+    console.log(`token//////// ${token}`);
+
+    const user = await this.userService.getUserInfoBysocialAccountUid(
+      req.user.kakaoId,
+    );
+
+    return {
+      token: token,
+      user: user,
+    };
+  }
+
+  // -----소셜로그인
+
   // 닉네임 중복 체크 : O
+
   @Get('/check/:nickname')
   async getCheckNicknameOverlap(
     @Param('nickname') nickname: string,
